diabetic_retinopathy_project/requirements.txt
flask
tensorflow
numpy
pillow
pandas
scikit-learn
opencv-python
